fix: move createRoute to lib as it's not really a service
refs #1701
