+++ apps/api/src/address/routes/address/address.router.ts
@@ -7,7 +7,7 @@ import {
   GetAddressTransactionsParamsSchema,
   GetAddressTransactionsResponseSchema
 } from "@src/address/http-schemas/address.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
