+++ apps/api/src/auth/routes/api-keys/api-keys.router.ts
@@ -10,7 +10,7 @@ import {
   UpdateApiKeyRequestSchema
 } from "@src/auth/http-schemas/api-key.schema";
 import { ListApiKeysResponseSchema } from "@src/auth/http-schemas/api-key.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_BEARER_OR_API_KEY } from "@src/core/services/openapi-docs/openapi-security";
