+++ apps/api/src/auth/routes/send-verification-email/send-verification-email.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 import { z } from "zod";
 
 import { AuthController } from "@src/auth/controllers/auth/auth.controller";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
