+++ apps/api/src/auth/routes/verify-email/verify-email.router.ts
@@ -3,7 +3,7 @@ import { container } from "tsyringe";
 
 import { AuthController } from "@src/auth/controllers/auth/auth.controller";
 import { VerifyEmailRequestSchema } from "@src/auth/http-schemas/verify-email.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
