+++ apps/api/src/bid/routes/bids/bids.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 
 import { BidController } from "@src/bid/controllers/bid/bid.controller";
 import { ListBidsQuerySchema, ListBidsResponseSchema } from "@src/bid/http-schemas/bid.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_BEARER_OR_API_KEY } from "@src/core/services/openapi-docs/openapi-security";
