+++ apps/api/src/billing/routes/get-balances/get-balances.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 
 import { WalletController } from "@src/billing/controllers/wallet/wallet.controller";
 import { GetBalancesQuerySchema, GetBalancesResponseOutputSchema } from "@src/billing/http-schemas/balance.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
