+++ apps/api/src/billing/routes/stripe-prices/stripe-prices.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 import { z } from "zod";
 
 import { StripeController } from "@src/billing/controllers/stripe/stripe.controller";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_BEARER_OR_API_KEY } from "@src/core/services/openapi-docs/openapi-security";
