+++ apps/api/src/billing/routes/stripe-webhook/stripe-webhook.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 import { z } from "zod";
 
 import { CheckoutController } from "@src/billing/controllers/checkout/checkout.controller";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
