+++ apps/api/src/billing/routes/usage/usage.router.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 
 import { UsageController } from "@src/billing/controllers/usage/usage.controller";
 import { GetUsageHistoryQuerySchema, UsageHistoryResponseSchema, UsageHistoryStatsResponseSchema } from "@src/billing/http-schemas/usage.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
