+++ apps/api/src/block/routes/block-prediction/block-prediction.router.ts
@@ -8,7 +8,7 @@ import {
   GetPredictedDateHeightResponseSchema,
   GetPredictionQuerySchema
 } from "@src/block/http-schemas/block-prediction.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
