+++ apps/api/src/block/routes/blocks/blocks.router.ts
@@ -7,7 +7,7 @@ import {
   ListBlocksQuerySchema,
   ListBlocksResponseSchema
 } from "@src/block/http-schemas/block.schema";
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
