+++ apps/api/src/certificate/routes/certificate.router.ts
@@ -1,6 +1,6 @@
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_BEARER_OR_API_KEY } from "@src/core/services/openapi-docs/openapi-security";
 import { CertificateController } from "../controllers/certificate.controller";
