+++ apps/api/src/dashboard/routes/graph-data/graph-data.router.ts
@@ -1,7 +1,7 @@
 import { LoggerService } from "@akashnetwork/logging";
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
 import { GraphDataController } from "@src/dashboard/controllers/graph-data/graph-data.controller";
