+++ apps/api/src/dashboard/routes/market-data/market-data.router.ts
@@ -1,6 +1,6 @@
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
 import { MarketDataController } from "@src/dashboard/controllers/market-data/market-data.controller";
