+++ apps/api/src/provider/routes/provider-earnings/provider-earnings.router.ts
@@ -1,7 +1,7 @@
 import type { Context } from "hono";
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
 import { ProviderEarningsController } from "@src/provider/controllers/provider-earnings/provider-earnings.controller";
