+++ apps/api/src/template/routes/templates/templates.router.ts
@@ -1,6 +1,6 @@
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
 import { TemplateController } from "@src/template/controllers/template/template.controller";
