+++ apps/api/src/user/routes/create-anonymous-user/create-anonymous-user.router.ts
@@ -1,7 +1,7 @@
 import { OpenAPIHono } from "@hono/zod-openapi";
 import { container } from "tsyringe";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { SECURITY_NONE } from "@src/core/services/openapi-docs/openapi-security";
 import type { ClientInfoContextVariables } from "@src/middlewares/clientInfoMiddleware";
 import { UserController } from "@src/user/controllers/user/user.controller";
