+++ apps/api/src/user/routes/get-anonymous-user/get-anonymous-user.router.ts
@@ -1,7 +1,7 @@
 import { container } from "tsyringe";
 import { z } from "zod";
 
-import { createRoute } from "@src/core/services/create-route/create-route";
+import { createRoute } from "@src/core/lib/create-route/create-route";
 import { OpenApiHonoHandler } from "@src/core/services/open-api-hono-handler/open-api-hono-handler";
 import { SECURITY_BEARER_OR_API_KEY } from "@src/core/services/openapi-docs/openapi-security";
 import { UserController } from "@src/user/controllers/user/user.controller";
